Sisco/improvements (#22)
* docs: initial project todo

* [phase 1.1] Pre-compute Comment Style Flags

* docs: reformat project todo

* [phase 1.2] Add IFormattable to Token

* [phase 1.3] Fix AppendToBuffer Inefficiency

* [phase 1.4] Cache SiblingIndex in RedNode

* [phase 2.1] Replace List<char> with ArrayPoolBufferWriter

* [phase 2.2] Build Operator Trie

* [phase 2.3] Standardize Position Types

* [phase 2.4] Unify Operator Matching with OperatorTrie

* [phase 3.1] Consolidate Duplicate Parsing Logic

* [phase 3.2] Address Schema Nullability

* [phase 3.3] Make Trivia Types Public

* [phase 3.5] Remove Obsolete APIs

* [phase 4.1] Error Recovery Test Suite

* [phase 4.2] Unicode Test Suite

* [phase 4.3] Numeric Edge Cases Tests

* [phase 4.4] Thread Safety Tests

* chore: setup category traits for all tests

* [phase 4.5] Performance Benchmarks

* fix: test failures after rebase

* chore: consolidate common logic into TokenizerCore

+++ TinyTokenizer/SimpleToken.cs
@@ -12,11 +12,15 @@ namespace TinyTokenizer;
 /// </summary>
 /// <param name="Type">The type of this token.</param>
 /// <param name="Content">The character content of this token.</param>
-/// <param name="Position">The absolute position in the source where this token starts.</param>
+/// <param name="Position">The absolute position in the source where this token starts. Limited to ~2GB file size.</param>
+/// <remarks>
+/// Position uses <c>int</c> rather than <c>long</c> for memory efficiency and alignment with
+/// .NET's string and array length limits. This supports files up to ~2GB in size.
+/// </remarks>
 public readonly record struct SimpleToken(
     SimpleTokenType Type,
     ReadOnlyMemory<char> Content,
-    long Position
+    int Position
 ) : ITextSerializable,
     IParsable<SimpleToken>,
     ISpanParsable<SimpleToken>,
@@ -53,7 +57,7 @@ public readonly record struct SimpleToken(
     /// <summary>
     /// Creates a simple token from a single character.
     /// </summary>
-    public static SimpleToken FromChar(SimpleTokenType type, char c, long position)
+    public static SimpleToken FromChar(SimpleTokenType type, char c, int position)
     {
         return new SimpleToken(type, new[] { c }.AsMemory(), position);
     }
@@ -61,7 +65,7 @@ public readonly record struct SimpleToken(
     /// <summary>
     /// Creates a simple token from a string.
     /// </summary>
-    public static SimpleToken FromString(SimpleTokenType type, string content, long position)
+    public static SimpleToken FromString(SimpleTokenType type, string content, int position)
     {
         return new SimpleToken(type, content.AsMemory(), position);
     }
@@ -144,7 +148,7 @@ public readonly record struct SimpleToken(
 
         // Parse position
         var positionSpan = s[(atIndex + 1)..];
-        if (!long.TryParse(positionSpan, NumberStyles.Integer, provider, out var position))
+        if (!int.TryParse(positionSpan, NumberStyles.Integer, provider, out var position))
             return false;
 
         // Extract content (between colon and @)
